# Applied perfomance fix for CUDA