This PR introduces a performance optimization fix for ONNX models running with CUDAExecutionProvider. The improvement is based on insights from the debugging methodology outlined in [this article](https://medium.com/neuml/debug-onnx-gpu-performance-c9290fe07459).
I have close to zero knowledge when it comes to ONNX models and optimization but I managed to speed up inference substantially.

### Findings
Benchmarks were run before and after the fix using two different test texts, with performance measured in seconds. The fix results in a significant reduction in inference time, particularly for longer inputs. 
**I also noticed a heavy VRAM reduction from about 2.2Gb to less than 750Mb**
#### Test 1
Input text:
```
You can now select additional permissions when creating an API key to use in any third-party libraries or software that integrate with Immich. This mechanism will give you better control over what the other applications or libraries can do with your Immich’s instance.
```

Run | Before Fix (s) | After Fix (s)
-- | -- | --
1 | 1.25 | 1.27
2 | 1.23 | 1.07
3 | 1.19 | 0.98
4 | 1.18 | 0.98
5 | 1.19 | 1.00

#### Test 2
Input text:
```
I figured that surely, someone has had this idea and built it before. On eBay you'll find cubes of resin embedding various random components from mechanical watches, but they are typically sold as "steampunk art" and bear little resemblance to the proper assembly of a mechanical watch movement. Sometimes, you'll find resin castings showing every component of a movement spread out in a plane like a buffet---very cool, but not what I'm looking for. Despite my best efforts, I haven't found anyone who makes what I'm after, and I have a sneaking suspicion as to why that is. 
Building an exploded view of a mechanical watch movement is undoubtedly very fiddly work and requires working knowledge about how a mechanical watch is assembled. People with that skillset are called watchmakers. Maker, not "destroyer for the sake of art". I guess it falls to me, then, to give this project an honest shot. 
```

Run | Before Fix | After Fix
-- | -- | --
1 | 21.79s | 3.21s
2 | 22.07s | 2.97s
3 | 22.48s | 2.88s
4 | 22.07s | 2.89s
5 | 22.72s | 2.89s

### Regressions observed
No noticeable issues anywhere except these warnings:
```javascript
2025-06-22 22:00:30.655884509 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.2/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.008611780 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.179631203 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.520232414 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_convs.1/Conv) running in Fallback mode. May be extremely slow.
```


<details>
  <summary>Full logs of warnings</summary>
  
  ```javascript
2025-06-22 22:00:30.624172353 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.0/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.624651230 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.0/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.626567581 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.1/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.626855339 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.1/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.655460468 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.1/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.655884509 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.2/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.656239449 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.2/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:30.656457937 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0_proj/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.007158269 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_convs.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.007646875 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.007949742 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.008198745 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.008400653 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.008611780 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.008805687 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.0/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.009298610 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.0/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.009528208 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.0/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.011287437 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.1/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.011502872 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.1/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.095264005 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.1/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.095733925 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.2/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.096008410 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N.2/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.096211064 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/N_proj/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.096619706 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/N_conv/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.096945494 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/F0_conv/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.097300665 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/text_encoder/cnn.0/cnn.0.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.097493754 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/text_encoder/cnn.1/cnn.1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.097679899 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/text_encoder/cnn.2/cnn.2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.102252449 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/encode/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.102503449 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/encode/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.102684659 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/encode/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.102847876 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/asr_res/asr_res.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.103066629 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.0/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.103271261 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.0/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.103439821 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.0/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.103640204 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.1/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.103846716 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.1/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.104071686 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.1/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.104274252 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.2/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.104477804 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.2/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.104645491 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.2/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.108035025 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.3/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.108382667 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.3/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.178637825 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/decode.3/conv1x1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.179391834 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.179631203 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.179851802 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.180051151 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.180285278 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.180478579 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.1/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.180847121 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.0/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.181087889 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.0/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.181308494 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.0/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.181511135 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.0/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.181716228 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.0/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.181918276 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.0/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.182207675 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.2/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.182428568 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.2/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.182642262 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.2/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.182843051 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.2/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.183049503 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.2/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.183239135 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.2/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.520232414 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_convs.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.520681323 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.1/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.521018531 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.1/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.521261891 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.1/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.521544489 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.1/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.521772198 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.1/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.522028272 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/noise_res.1/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.522248235 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.4/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.522457247 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.4/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.522666828 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.4/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.522871930 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.4/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.523089521 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.4/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.523286211 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.4/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.523602816 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.3/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.523829089 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.3/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.524044914 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.3/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.524240359 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.3/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.524448575 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.3/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.524640878 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.3/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.524929090 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.5/convs1.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.525146196 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.5/convs2.0/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.525358021 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.5/convs1.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.525613007 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.5/convs2.1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.525836700 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.5/convs1.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.526079740 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/resblocks.5/convs2.2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:31.526258121 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/decoder/decoder/generator/conv_post/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:32.176433063 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.0/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:32.176936786 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.0/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:32.178846164 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.1/conv1/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:32.179079388 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.1/conv2/Conv) running in Fallback mode. May be extremely slow.
2025-06-22 22:00:32.201871486 [W:onnxruntime:Default, conv.cc:425 UpdateState] OP Conv(/encoder/F0.1/conv1x1/Conv) running in Fallback mode. May be extremely slow.
…