@M4TH1EU 2025-06-24
Hi,

I've ran the test script with and without the fix and below are the results.
Strangely the VRAM usage is different than what I said previously, it instead increases by 1Gb but also fastens inference by 6-7x.

## with fix

```bash
(sources) root@ai:~/speaches/sources# python scripts/tts_performance_test.py 
2025-06-24 11:31:50,433:DEBUG:__main__:main:54:Config: {"api_key":"**********","log_level":"debug","logs_directory":"logs","speaches_base_url":"**********","speech_model_id":"speaches-ai/Kokoro-82M-v1.0-ONNX","voice_id":"af_heart","input_text":"You can now select additional permissions when creating an API key to use in any third-party libraries or software that integrate with Immich. This mechanism will give you better control over what the other applications or libraries can do with your Immich’s instance.","iterations":5,"concurrency":1}
2025-06-24 11:31:50,449:DEBUG:__main__:main:65:Attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX
2025-06-24 11:31:50,455:INFO:__main__:main:67:Finished attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX. Response: {"detail":"Not Found"}
2025-06-24 11:31:51,833:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.1747058620094322, 'average_chunk_time': 0.39161954100321356, 'total_chunks': 3, 'total_time': 1.174982477008598}
2025-06-24 11:31:52,954:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.1180327649926767, 'average_chunk_time': 0.3727274479994473, 'total_chunks': 3, 'total_time': 1.1183128119882895}
2025-06-24 11:31:54,057:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.0999500360048842, 'average_chunk_time': 0.36669327633959864, 'total_chunks': 3, 'total_time': 1.1002879110019421}
2025-06-24 11:31:55,167:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.1068688040104462, 'average_chunk_time': 0.3689947533372712, 'total_chunks': 3, 'total_time': 1.1072000730055152}
2025-06-24 11:31:56,273:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.1018338139983825, 'average_chunk_time': 0.36731715400431614, 'total_chunks': 3, 'total_time': 1.1021541749942116}
2025-06-24 11:31:56,273:INFO:__main__:main:104:All tasks completed in 5.62 seconds

(sources) root@ai:~/speaches/sources# python scripts/tts_performance_test.py 
2025-06-24 11:31:09,059:DEBUG:__main__:main:54:Config: {"api_key":"**********","log_level":"debug","logs_directory":"logs","speaches_base_url":"**********","speech_model_id":"speaches-ai/Kokoro-82M-v1.0-ONNX","voice_id":"af_heart","input_text":"I figured that surely, someone has had this idea and built it before. On eBay you'll find cubes of resin embedding various random components from mechanical watches, but they are typically sold as \"steampunk art\" and bear little resemblance to the proper assembly of a mechanical watch movement. Sometimes, you'll find resin castings showing every component of a movement spread out in a plane like a buffet---very cool, but not what I'm looking for. Despite my best efforts, I haven't found anyone who makes what I'm after, and I have a sneaking suspicion as to why that is. Building an exploded view of a mechanical watch movement is undoubtedly very fiddly work and requires working knowledge about how a mechanical watch is assembled. People with that skillset are called watchmakers. Maker, not \"destroyer for the sake of art\". I guess it falls to me, then, to give this project an honest shot. ","iterations":5,"concurrency":1}
2025-06-24 11:31:09,075:DEBUG:__main__:main:65:Attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX
2025-06-24 11:31:09,081:INFO:__main__:main:67:Finished attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX. Response: {"detail":"Not Found"}
2025-06-24 11:31:12,354:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.7301472699909937, 'average_chunk_time': 0.38530315500065626, 'total_chunks': 8, 'total_time': 3.0826160819997313}
2025-06-24 11:31:15,404:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.6916778290033108, 'average_chunk_time': 0.38084321549831657, 'total_chunks': 8, 'total_time': 3.0470471610024106}
2025-06-24 11:31:18,457:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.7069427340029506, 'average_chunk_time': 0.38120678975246847, 'total_chunks': 8, 'total_time': 3.0498365300009027}
2025-06-24 11:31:21,489:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.6775393890129635, 'average_chunk_time': 0.37867880474914273, 'total_chunks': 8, 'total_time': 3.029638548992807}
2025-06-24 11:31:24,533:DEBUG:__main__:create_speech:94:{'time_to_first_token': 1.6811024449998513, 'average_chunk_time': 0.3800435333760106, 'total_chunks': 8, 'total_time': 3.040532286002417}
2025-06-24 11:31:24,533:INFO:__main__:main:104:All tasks completed in 15.26 seconds
```

```bash
root@pve:~# nvidia-smi
Tue Jun 24 11:32:35 2025       
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 570.169                Driver Version: 570.169        CUDA Version: 12.8     |
|-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  Quadro P4000                   Off |   00000000:01:00.0 Off |                  N/A |
| 48%   45C    P8              6W /  105W |    5531MiB /   8192MiB |      0%      Default |
|                                         |                        |                  N/A |
+-----------------------------------------+------------------------+----------------------+
                                                                                         
+-----------------------------------------------------------------------------------------+
| Processes:                                                                              |
|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
|        ID   ID                                                               Usage      |
|=========================================================================================|
|    0   N/A  N/A          901006      C   ...ma.cpp/build/bin/llama-server        100MiB |
|    0   N/A  N/A          901009      C   ...ma.cpp/build/bin/llama-server       4292MiB |
|    0   N/A  N/A         2730227      C   ...hes/sources/.venv/bin/python3       1136MiB | <-- speaches
+-----------------------------------------------------------------------------------------+
```

## without fix

```bash
(sources) root@ai:~/speaches/sources# python scripts/tts_performance_test.py 
2025-06-24 11:34:18,227:DEBUG:__main__:main:54:Config: {"api_key":"**********","log_level":"debug","logs_directory":"logs","speaches_base_url":"**********","speech_model_id":"speaches-ai/Kokoro-82M-v1.0-ONNX","voice_id":"af_heart","input_text":"You can now select additional permissions when creating an API key to use in any third-party libraries or software that integrate with Immich. This mechanism will give you better control over what the other applications or libraries can do with your Immich’s instance.","iterations":5,"concurrency":1}
2025-06-24 11:34:18,243:DEBUG:__main__:main:65:Attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX
2025-06-24 11:34:18,248:INFO:__main__:main:67:Finished attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX. Response: {"detail":"Not Found"}
2025-06-24 11:34:26,867:DEBUG:__main__:create_speech:94:{'time_to_first_token': 7.1291544130072, 'average_chunk_time': 2.3764389813344073, 'total_chunks': 3, 'total_time': 7.1294396789890015}
2025-06-24 11:34:34,339:DEBUG:__main__:create_speech:94:{'time_to_first_token': 7.469057883005007, 'average_chunk_time': 2.4897439390042564, 'total_chunks': 3, 'total_time': 7.469370872990112}
2025-06-24 11:34:41,982:DEBUG:__main__:create_speech:94:{'time_to_first_token': 7.639139992999844, 'average_chunk_time': 2.5464232506686435, 'total_chunks': 3, 'total_time': 7.6394611230061855}
2025-06-24 11:34:49,656:DEBUG:__main__:create_speech:94:{'time_to_first_token': 7.670357412993326, 'average_chunk_time': 2.5568285739961234, 'total_chunks': 3, 'total_time': 7.67067757199402}
2025-06-24 11:34:57,427:DEBUG:__main__:create_speech:94:{'time_to_first_token': 7.768139161009458, 'average_chunk_time': 2.5894171756663127, 'total_chunks': 3, 'total_time': 7.768456675999914}
2025-06-24 11:34:57,427:INFO:__main__:main:104:All tasks completed in 37.70 seconds

(sources) root@ai:~/speaches/sources# python scripts/tts_performance_test.py 
2025-06-24 11:35:27,586:DEBUG:__main__:main:54:Config: {"api_key":"**********","log_level":"debug","logs_directory":"logs","speaches_base_url":"**********","speech_model_id":"speaches-ai/Kokoro-82M-v1.0-ONNX","voice_id":"af_heart","input_text":"I figured that surely, someone has had this idea and built it before. On eBay you'll find cubes of resin embedding various random components from mechanical watches, but they are typically sold as \"steampunk art\" and bear little resemblance to the proper assembly of a mechanical watch movement. Sometimes, you'll find resin castings showing every component of a movement spread out in a plane like a buffet---very cool, but not what I'm looking for. Despite my best efforts, I haven't found anyone who makes what I'm after, and I have a sneaking suspicion as to why that is. Building an exploded view of a mechanical watch movement is undoubtedly very fiddly work and requires working knowledge about how a mechanical watch is assembled. People with that skillset are called watchmakers. Maker, not \"destroyer for the sake of art\". I guess it falls to me, then, to give this project an honest shot. ","iterations":5,"concurrency":1}
2025-06-24 11:35:27,602:DEBUG:__main__:main:65:Attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX
2025-06-24 11:35:27,608:INFO:__main__:main:67:Finished attempting to pull model speaches-ai/Kokoro-82M-v1.0-ONNX. Response: {"detail":"Not Found"}
2025-06-24 11:35:49,616:DEBUG:__main__:create_speech:94:{'time_to_first_token': 10.853605080003035, 'average_chunk_time': 2.6836399167514173, 'total_chunks': 8, 'total_time': 21.46932343399385}
2025-06-24 11:36:12,110:DEBUG:__main__:create_speech:94:{'time_to_first_token': 11.80665492398839, 'average_chunk_time': 2.811301073874347, 'total_chunks': 8, 'total_time': 22.490732182996}
2025-06-24 11:36:34,683:DEBUG:__main__:create_speech:94:{'time_to_first_token': 11.88121530499484, 'average_chunk_time': 2.8212509331260662, 'total_chunks': 8, 'total_time': 22.57018496299861}
2025-06-24 11:36:57,506:DEBUG:__main__:create_speech:94:{'time_to_first_token': 12.054876244990737, 'average_chunk_time': 2.852417277126733, 'total_chunks': 8, 'total_time': 22.819525539991446}
2025-06-24 11:37:20,514:DEBUG:__main__:create_speech:94:{'time_to_first_token': 12.177433370001381, 'average_chunk_time': 2.875627628627626, 'total_chunks': 8, 'total_time': 23.00520208099624}
2025-06-24 11:37:20,514:INFO:__main__:main:104:All tasks completed in 112.38 seconds
```



```bash
root@pve:~# nvidia-smi
Tue Jun 24 11:36:55 2025       
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 570.169                Driver Version: 570.169        CUDA Version: 12.8     |
|-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  Quadro P4000                   Off |   00000000:01:00.0 Off |                  N/A |
| 46%   34C    P8              5W /  105W |    4497MiB /   8192MiB |      0%      Default |
|                                         |                        |                  N/A |
+-----------------------------------------+------------------------+----------------------+
                                                                                         
+-----------------------------------------------------------------------------------------+
| Processes:                                                                              |
|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
|        ID   ID                                                               Usage      |
|=========================================================================================|
|    0   N/A  N/A          901006      C   ...ma.cpp/build/bin/llama-server        100MiB |
|    0   N/A  N/A          901009      C   ...ma.cpp/build/bin/llama-server       4292MiB |
|    0   N/A  N/A         2739767      C   ...hes/sources/.venv/bin/python3        102MiB | <-- speaches
+-----------------------------------------------------------------------------------------+
```